machine-learning-files/requirements.txt
fastapi==0.92.0
matplotlib==3.6.3
Pillow==9.4.0
torch==1.13.1
torchinfo==1.7.2
torchvision==0.14.1
uvicorn==0.20.0